pyproject.toml
[project]
name = "terraform-google-private-gke-cluster"
version = "0.0.1"
description = "Test harness for Terraform/Tofu module"
readme = "README.md"
requires-python = ">=3.12"
# spell-checker: disable
dependencies = [
    "cryptography>=46.0.1",
    "google-cloud-artifact-registry>=1.16.1",
    "google-cloud-compute>=1.37.0",
    "google-cloud-container>=2.59.0",
    "google-cloud-iam>=2.19.1",
    "google-cloud-resource-manager>=1.14.2",
    "kubernetes>=33.1.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pyright[nodejs]>=1.1.405",
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
ignore = []
select = ["ALL"]
extend-select = ["I"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Allow assert and don't complain about implicit namespaces or too many arguements for tests.
"**/tests/**/*" = ["S", "INP", "PLR0913"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.coverage.paths]
source = ["tests"]

[tool.coverage.report]
show_missing = true

[tool.coverage.run]
branch = true

[tool.pytest.ini_options]
addopts = "--import-mode=importlib --cov=tests --cov-report=term-missing"
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s - %(levelname)s:%(name)s:%s(message)"
log_cli_date_format = "%Y-%m-%dT%H:%M:%S%z"
testpaths = ["tests"]
